+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+* feat: persist `Meteor.users` collection by default so `Meteor.user()` works as expected when offline
+* feat: support `Mongo.ObjectID`
+* fix: under-the-hood optimizations
+
 ## 0.1.2
 * fix: error on initial load - `One of the specified object stores was not found`